requirements.txt
Kivy==2.0.0
pandas==1.1.4
Pillow==8.1.0
